Refactor ManagedLanding into ManagedLandingContent
This way, if the placeholder is loaded for whatever reason, the API
requests (to Monitors, Credentials, and Contacts) will not be made.
These will usually be oauth errors anyway, so it's good to prevent them.
